linux-t20: Fix issue with USB OTG on Acer
An issue was reported that the device was not able to setup a rndis
connection to a Acer, Aspire V15 Nitro, VN7-591G-70X0.

It worked if the VCB was connected trough a USB hub to the PC. Which
indicated that it was something with OTG functionality that produced
this error.

So we actually disable the OTG functionality completely and only enable
USB device functionality on that port. The OTG functionality was
unnecessary and unused anyway.

Fixed #16