linux-t20: colibri-t20: re-worked sja1000 platform map
We changed how SJA1000 is accessed in VCB P1E revision, what we did was
to use multiplexed access mode on the external memory controller. We
earlier did this in software. This change will decrease the CPU load for
SJA1000 interrupts by half!.

To be able to maintain same software image we need to be able to
dynamicly determain which CAN access mode to use.

This is controlled by a kernel argument:

vcb_muxed_can

So on revisions P1E and further we need to pass 'vcb_muxed_can=1' to the
kernel for it to use correct bus access method.

We also use the vcb_muxed_can switch to add a third CAN controller which
was added in P1E revison as well.